pyproject.toml
[tool.poetry]
name = "messenger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
black = "^25.1.0"
pre-commit = "^4.2.0"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
uvicorn = {extras = ["standart"], version = "^0.35.0"}
websockets = "^15.0.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.35.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
